Factor 'edit config' command out of editor and into configuration
