+++ browser/src/Editor/NeovimEditor/NeovimEditorCommands.ts
@@ -4,18 +4,14 @@
  * Contextual commands for NeovimEditor
  *
  */
-import * as fs from "fs"
 import * as os from "os"
-import * as path from "path"
 
 import { clipboard } from "electron"
 import * as Oni from "oni-api"
 
 import { NeovimInstance } from "./../../neovim"
 import { CallbackCommand, CommandManager } from "./../../Services/CommandManager"
-import { getUserConfigFilePath } from "./../../Services/Configuration"
 import { ContextMenuManager } from "./../../Services/ContextMenu"
-import { editorManager } from "./../../Services/EditorManager"
 import { findAllReferences, format, LanguageEditorIntegration } from "./../../Services/Language"
 import { MenuManager } from "./../../Services/Menu"
 import { QuickOpen } from "./../../Services/QuickOpen"
@@ -102,24 +98,6 @@ export class NeovimEditorCommands {
             await neovimInstance.command("set paste")
             await neovimInstance.input(sanitizedText)
             await neovimInstance.command("set nopaste")
-        }
-
-        const openDefaultConfig = async (): Promise<void> => {
-            const activeEditor = editorManager.activeEditor
-            const buf = await activeEditor.openFile(getUserConfigFilePath())
-            const lineCount = buf.lineCount
-
-            if (lineCount === 1) {
-                const defaultConfigJsPath = path.join(
-                    __dirname,
-                    "configuration",
-                    "config.default.js",
-                )
-                const defaultConfigLines = fs
-                    .readFileSync(defaultConfigJsPath, "utf8")
-                    .split(os.EOL)
-                await buf.setLines(0, defaultConfigLines.length, defaultConfigLines)
-            }
         }
 
         const shouldShowMenu = () => {
@@ -211,13 +189,6 @@ export class NeovimEditorCommands {
             new CallbackCommand("language.symbols.workspace", null, null, () =>
                 this._symbols.openWorkspaceSymbolsMenu(),
             ),
-            new CallbackCommand(
-                "oni.config.openConfigJs",
-                "Configuration: Edit Oni Config",
-                "Edit configuration file ('config.js') for Oni",
-                () => openDefaultConfig(),
-            ),
-
             new CallbackCommand(
                 "oni.config.openInitVim",
                 "Configuration: Edit Neovim Config",
